feat(scrolling): prevent section overscrolling
When a section with scroll overflow is preceeded and / or succeeded by
another section, the last scroll in the overflowed section also triggers
a scroll to the next (or previous) section. This piece of code authored
by @Paperweb and presented in a GitHub issue comment does just that
by setting a scroll time delay.